Move angular-b2c-sample-app and rename
