+++ samples/msal-angular-samples/angular-b2c-sample/jest.config.js
