+++ samples/msal-angular-samples/angular-b2c-sample/karma.conf.js
