+++ samples/msal-angular-samples/angular-b2c-sample/package.json
