+++ samples/msal-angular-samples/angular-b2c-sample/src/app/app-routing.module.ts
