+++ samples/msal-angular-samples/angular-b2c-sample/src/app/app.component.spec.ts
