+++ samples/msal-angular-samples/angular-b2c-sample/src/app/failed/failed.component.ts
