+++ samples/msal-angular-samples/angular-b2c-sample/src/app/home/home.component.html
