+++ samples/msal-angular-samples/angular-b2c-sample/src/app/home/home.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { AppModule } from './../app.module';
+import { AppModule } from '../app.module';
 
 import { HomeComponent } from './home.component';
