+++ samples/msal-angular-samples/angular-b2c-sample/src/app/profile/profile.component.ts
