+++ samples/msal-angular-samples/angular-b2c-sample/src/environments/environment.e2e.ts
