+++ samples/msal-angular-samples/angular-b2c-sample/src/environments/environment.ts
