+++ samples/msal-angular-samples/angular-b2c-sample/src/test.ts
