+++ samples/msal-angular-samples/angular-b2c-sample/test/local-account.spec.ts
