+++ samples/msal-angular-samples/angular-b2c-sample/test/msa-account.spec.ts
